requirements.txt
streamlit
tensorflow
scikit-learn
matplotlib
seaborn
opencv-python
numpy
pandas
Pillow